Create function modularized
