--- client/components/GameFunctions/addPlayers.js
@@ -1,28 +1,0 @@
-export function addPlayer(self, playerInfo) {
-  self.ship = self.physics.add
-    .image(playerInfo.x, playerInfo.y, 'ship')
-    .setOrigin(0.5, 0.5)
-    .setDisplaySize(53, 40)
-  if (playerInfo.team === 'blue') {
-    self.ship.setTint(0x0000ff)
-  } else {
-    self.ship.setTint(0xff0000)
-  }
-  self.ship.setDrag(100)
-  self.ship.setAngularDrag(100)
-  self.ship.setMaxVelocity(200)
-}
-
-export function addOtherPlayers(self, playerInfo) {
-  const otherPlayer = self.add
-    .sprite(playerInfo.x, playerInfo.y, 'otherPlayer')
-    .setOrigin(0.5, 0.5)
-    .setDisplaySize(53, 40)
-  if (playerInfo.team === 'blue') {
-    otherPlayer.setTint(0x0000ff)
-  } else {
-    otherPlayer.setTint(0xff0000)
-  }
-  otherPlayer.playerId = playerInfo.playerId
-  self.otherPlayers.add(otherPlayer)
-}
