+++ client/components/GameFunctions/createFunctions/index.js
@@ -1,0 +1,6 @@
+import { players } from './addPlayers'
+import { keyboardControls } from './keyboardControls'
+import { scoreAndStars } from './lifeStars';
+import { NPCCells } from './cells'
+
+export {players, keyboardControls, scoreAndStars, NPCCells}
